+++ src/client.rs
@@ -40,25 +40,37 @@ macro_rules! event_handler_function {
 }
 
 /// Wrapper around the [`JoinHandle`] returned by [`Client::start`]
+#[allow(clippy::module_name_repetitions)]
 pub struct ClientThread(JoinHandle<()>, Sender<()>);
 
 impl ClientThread {
-    /// Alias of [`JoinHandle::join`]
+    // Ignore missing error docs because it's an alias of `join`
+    #[allow(clippy::missing_errors_doc)]
+    /// Alias of [`JoinHandle::join()`]
     pub fn join(self) -> std::thread::Result<()> {
         self.0.join()
     }
 
+    // Ignore missing error docs because it's an alias of `is_finished`
+    #[allow(clippy::missing_errors_doc)]
+    #[must_use]
     /// Alias of [`JoinHandle::is_finished`]
     pub fn is_finished(&self) -> bool {
         self.0.is_finished()
     }
-
+    // Ignore missing error docs because it's an alias of `thread`
+    #[allow(clippy::missing_errors_doc)]
+    #[must_use]
     /// Alias of [`JoinHandle::thread`]
     pub fn thread(&self) -> &Thread {
         self.0.thread()
     }
 
     /// Attempt to stop the client's send and receive loop
+    ///
+    /// # Errors
+    /// - Failed to send stop message (maybe it has already stopped?)
+    /// - The event loop had its own error
     pub fn stop(self) -> Result<()> {
         // Attempt to send the message to stop the thread
         self.1.send(())?;
@@ -67,13 +79,18 @@ impl ClientThread {
 
         Ok(())
     }
+
+    /// "Forgets" client thread, removing the variable, but keeping the client running indefinitely.
+    pub fn persist(self) {
+        std::mem::forget(self);
+    }
 }
 
 /// The Discord client
-#[derive(Clone)]
 pub struct Client {
     connection_manager: ConnectionManager,
     event_handler_registry: Arc<HandlerRegistry>,
+    thread: Option<ClientThread>,
 }
 
 #[cfg(feature = "bevy")]
@@ -81,6 +98,7 @@ impl bevy::ecs::system::Resource for Client {}
 
 impl Client {
     /// Creates a new `Client`
+    #[must_use]
     pub fn new(client_id: u64) -> Self {
         let event_handler_registry = Arc::new(HandlerRegistry::new());
         let connection_manager = ConnectionManager::new(client_id, event_handler_registry.clone());
@@ -88,16 +106,17 @@ impl Client {
         Self {
             connection_manager,
             event_handler_registry,
+            thread: None,
         }
     }
 
+    // TODO: Add examples
     /// Start the connection manager
     ///
     /// Only join the thread if there is no other task keeping the program alive.
     ///
     /// This must be called before all and any actions such as `set_activity`
-    #[must_use]
-    pub fn start(&mut self) -> ClientThread {
+    pub fn start(&mut self) {
         let (tx, rx) = crossbeam_channel::bounded::<()>(1);
 
         let thread = self.connection_manager.start(rx);
@@ -110,7 +129,47 @@ impl Client {
         });
 
         forget(ready);
-        ClientThread(thread, tx)
+        self.thread = Some(ClientThread(thread, tx));
+    }
+
+    /// Shutdown the client and its thread
+    ///
+    /// # Errors
+    /// - The internal connection thread ran into an error
+    /// - The client was not started, or has already been shutdown
+    pub fn shutdown(self) -> Result<()> {
+        crate::READY.store(false, Ordering::Relaxed);
+        let Self { thread, .. } = self;
+
+        if let Some(thread) = thread {
+            thread.1.send(())?;
+            thread.join().map_err(|_| DiscordError::ThreadError)?;
+
+            Ok(())
+        } else {
+            Err(DiscordError::NotStarted)
+        }
+    }
+
+    /// Block indefinitely until the client shuts down
+    ///
+    /// This is nearly the same as [`Client::shutdown()`],
+    /// except that it does not attempt to stop the internal thread,
+    /// and rather waits for it to finish, which could never happen.
+    ///
+    /// # Errors
+    /// - The internal connection thread ran into an error
+    /// - The client was not started, or has already been shutdown
+    pub fn block_on(self) -> Result<()> {
+        let Self { thread, .. } = self;
+
+        if let Some(thread) = thread {
+            thread.join().map_err(|_| DiscordError::ThreadError)?;
+
+            Ok(())
+        } else {
+            Err(DiscordError::NotStarted)
+        }
     }
 
     /// Check if the client is ready
@@ -118,17 +177,12 @@ impl Client {
         crate::READY.load(Ordering::Acquire)
     }
 
-    /// Check if the client has started
-    pub fn is_started() -> bool {
-        crate::STARTED.load(Ordering::Acquire)
-    }
-
     fn execute<A, E>(&mut self, cmd: Command, args: A, evt: Option<Event>) -> Result<Payload<E>>
     where
         A: Serialize + Send + Sync,
         E: Serialize + DeserializeOwned + Send + Sync,
     {
-        if !crate::STARTED.load(Ordering::Relaxed) || !crate::READY.load(Ordering::Relaxed) {
+        if !crate::READY.load(Ordering::Relaxed) {
             return Err(DiscordError::NotStarted);
         }
 
@@ -149,6 +203,9 @@ impl Client {
     }
 
     /// Set the users current activity
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn set_activity<F>(&mut self, f: F) -> Result<Payload<Activity>>
     where
         F: FnOnce(Activity) -> Activity,
@@ -157,6 +214,9 @@ impl Client {
     }
 
     /// Clear the users current activity
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn clear_activity(&mut self) -> Result<Payload<Activity>> {
         self.execute(Command::SetActivity, SetActivityArgs::default(), None)
     }
@@ -165,6 +225,9 @@ impl Client {
     //       SEND_ACTIVITY_JOIN_INVITE and CLOSE_ACTIVITY_REQUEST are,
     //       they are not documented.
     /// Send an invite to a user to join a game
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn send_activity_join_invite(&mut self, user_id: u64) -> Result<Payload<Value>> {
         self.execute(
             Command::SendActivityJoinInvite,
@@ -174,6 +237,9 @@ impl Client {
     }
 
     /// Close request to join a game
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn close_activity_request(&mut self, user_id: u64) -> Result<Payload<Value>> {
         self.execute(
             Command::CloseActivityRequest,
@@ -183,6 +249,9 @@ impl Client {
     }
 
     /// Subscribe to a given event
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn subscribe<F>(&mut self, evt: Event, f: F) -> Result<Payload<Subscription>>
     where
         F: FnOnce(SubscriptionArgs) -> SubscriptionArgs,
@@ -191,6 +260,9 @@ impl Client {
     }
 
     /// Unsubscribe from a given event
+    ///
+    /// # Errors
+    /// - See [`DiscordError`] for more info
     pub fn unsubscribe<F>(&mut self, evt: Event, f: F) -> Result<Payload<Subscription>>
     where
         F: FnOnce(SubscriptionArgs) -> SubscriptionArgs,
@@ -271,20 +343,26 @@ impl Client {
     ///
     /// NOTE: Please only use this for the ready event, or if you know what you are doing.
     ///
-    /// # Panics
+    /// # Errors
+    /// - Channel disconnected
     ///
-    /// Panics if the channel is disconnected for whatever reason.
+    /// # Panics
+    /// - Panics if the channel is disconnected for whatever reason.
     pub fn block_until_event(&mut self, event: Event) -> Result<crate::event_handler::Context> {
-        let (tx, rx) = crossbeam_channel::bounded::<crate::event_handler::Context>(1);
+        let (tx, rx) = crossbeam_channel::unbounded::<crate::event_handler::Context>();
 
         let handler = move |info| {
-            dbg!(tx.send(info).err());
+            if let Err(e) = tx.send(info) {
+                error!("{e}");
+            }
         };
 
         // `handler` is automatically unregistered once this variable drops
         let _cb_handle = self.on_event(event, handler);
 
-        Ok(rx.recv()?)
+        let response = rx.recv()?;
+
+        Ok(response)
     }
 
     event_handler_function!(on_ready, Event::Ready);
@@ -301,15 +379,6 @@ impl Client {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_is_started() {
-        assert!(!Client::is_started());
-
-        crate::STARTED.store(true, Ordering::Relaxed);
-
-        assert!(Client::is_started());
-    }
 
     #[test]
     fn test_is_ready() {