+++ src/error.rs
@@ -1,54 +1,61 @@
-use crossbeam_channel::{RecvError, RecvTimeoutError, SendError};
+use crossbeam_channel::{RecvTimeoutError, SendError};
 use serde_json::Error as JsonError;
 use std::{
-    io::Error as IoError, result::Result as StdResult,
-    sync::mpsc::RecvTimeoutError as ChannelTimeout,
+    io::Error as IoError,
+    result::Result as StdResult,
+    sync::mpsc::{RecvError as ChannelRecv, RecvTimeoutError as ChannelTimeout},
 };
-use thiserror::Error as AsError;
 
 use crate::models::Message;
 
 /// Error types from Discord
-#[derive(Debug, AsError)]
+#[derive(Debug, thiserror::Error)]
+#[allow(clippy::module_name_repetitions)]
 pub enum DiscordError {
-    /// Io Error
     #[error("Io Error")]
+    /// Io Error
     IoError(#[from] IoError),
-    /// tx.send returned error
     #[error("Could not send message: {0}")]
-    SendMessage(#[from] SendError<Message>),
     /// tx.send returned error
+    SendMessage(#[from] SendError<Message>),
     #[error("Could not close event loop: {0}")]
+    /// tx.send returned error
     CloseError(#[from] SendError<()>),
-    /// Error Receiving message
     #[error("Error Receiving message")]
-    ReceiveError(#[from] RecvError),
-    /// Json Error
-    #[error("Error parsing Json")]
-    JsonError(#[from] JsonError),
-    /// Timeout Error
+    /// Error Receiving message
+    ReceiveError(#[from] crossbeam_channel::RecvError),
+    #[error("Error Receiving message")]
+    /// Error Receiving message
+    MPSCReceiveError(#[from] ChannelRecv),
     #[error("Error on Channel Timeout")]
-    Timeout(#[from] ChannelTimeout),
-    /// Receiving timed out
+    /// Timeout Error
+    MPSCTimeout(#[from] ChannelTimeout),
     #[error("Receiving timed out")]
-    RecvTimeoutError(#[from] RecvTimeoutError),
-    /// Option unwrapped to None
+    /// Receiving timed out
+    TimeoutError(#[from] RecvTimeoutError),
+    #[error("Error parsing Json")]
+    /// Json Error
+    JsonError(#[from] JsonError),
+    #[error("A thread ran into an error. See logs for more info.")]
+    /// A thread ran into an error
+    ThreadError,
     #[error("{0}")]
+    /// Option unwrapped to None
     NoneError(String),
-    /// Conversion Error
     #[error("Error converting values")]
+    /// Conversion Error
     Conversion,
-    /// Subscription Joining Error
     #[error("Error subscribing to an event")]
+    /// Subscription Joining Error
     SubscriptionFailed,
-    /// Connection Closing error
     #[error("Connection was closed prematurely")]
+    /// Connection Closing error
     ConnectionClosed,
-    /// Connection has not been started
     #[error("Connection has not been started")]
+    /// Connection has not been started
     NotStarted,
-    /// The send & receive loop ran into an error
     #[error("Event loop ran into an unknown error")]
+    /// The send & receive loop ran into an error
     EventLoopError,
     /// No changes were made to the event handler
     #[error("No changes were made to the event handler. This can usually be ignored")]
@@ -56,10 +63,20 @@ pub enum DiscordError {
 }
 
 impl DiscordError {
+    #[must_use]
     /// Tell whether an [`IoError`] would block the connection
     pub fn io_would_block(&self) -> bool {
         match self {
             Self::IoError(ref err) => err.kind() == std::io::ErrorKind::WouldBlock,
+            _ => false,
+        }
+    }
+
+    #[must_use]
+    /// Checks if the error should break the connection
+    pub fn should_break(&self) -> bool {
+        match self {
+            Self::IoError(ref err) => err.kind() == std::io::ErrorKind::ConnectionRefused,
             _ => false,
         }
     }