+++ src/lib.rs
@@ -1,9 +1,14 @@
-#![warn(missing_docs, rust_2018_compatibility, rust_2018_idioms, clippy::all)]
+#![warn(
+    missing_docs,
+    rust_2018_compatibility,
+    rust_2018_idioms,
+    clippy::all,
+    clippy::pedantic
+)]
 #![forbid(unsafe_code)]
 
 //! A Rust library that allows the developer to interact with the Discord Presence API with ease
 
-pub(crate) static STARTED: AtomicBool = AtomicBool::new(false);
 pub(crate) static READY: AtomicBool = AtomicBool::new(false);
 
 // Cannot remove this *macro_use*, would break derive inside of macros