+++ src/macros.rs
@@ -11,6 +11,7 @@ macro_rules! builder_func_doc {
 macro_rules! builder_func {
     [ $name:ident, $type:tt func ] => {
         #[doc = builder_func_doc!($type)]
+        #[must_use]
         pub fn $name<F>(mut self, func: F) -> Self
             where F: FnOnce($type) -> $type
         {
@@ -20,6 +21,7 @@ macro_rules! builder_func {
 
     [ $name:ident, String ] => {
         #[doc = builder_func_doc!(Stringish)]
+        #[must_use]
         pub fn $name<S>(mut self, value: S) -> Self
             where S: Into<String>
         {
@@ -29,6 +31,7 @@ macro_rules! builder_func {
 
     [ $name:ident, $type:ty ] => {
         #[doc = builder_func_doc!($type)]
+        #[must_use]
         pub fn $name(mut self, value: $type) -> Self {
             self.$name = Some(value); self
         }
@@ -96,6 +99,7 @@ macro_rules! builder {
     [ @im ( $name:ident ) -> ( $($out:tt)* ) ] => {
         impl $name {
             #[doc = concat!("Instantiates the `", stringify!($name), "` struct using the `Default` implementation")]
+            #[must_use]
             pub fn new() -> Self {
                 Self::default()
             }