+++ CHANGELOG.md
@@ -9,6 +9,14 @@ I will add links where possible, including retroactively if possible.
 
 ## [Unreleased](https://github.com/jewlexx/discord-presence/tree/trunk)
 
+## [0.6.0]
+
+### Breaking Changes
+
+- Send & Receive Loop now breaks for `ConnectionRefused` error kind, rather than `WouldBlock`
+- Removed client thread handle (now is kept internally on the Client struct)
+- Removed `STARTED` boolean. (Pretty much pointless as it is only different between when the client has been started, but is not yet ready)
+
 ### Fixed
 
 - Added back list of events for Bevy crate